+++ src/Feed/SourceType/Brnd/ApiClient.php
@@ -9,7 +9,6 @@ use App\Feed\BrndFeedType;
 use Psr\Cache\CacheItemInterface;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
@@ -25,6 +24,7 @@ class ApiClient
     private array $apiClients = [];
 
     public function __construct(
+        private readonly HttpClientInterface $client,
         private readonly CacheItemPoolInterface $feedsCache,
         private readonly LoggerInterface $logger,
     ) {}
@@ -130,7 +130,8 @@ class ApiClient
         }
 
         $secrets = new SecretsDTO($feedSource);
-        $this->apiClients[$id] = HttpClient::createForBaseUri($secrets->apiBaseUri)->withOptions([
+        $this->apiClients[$id] = $this->client->withOptions([
+            'base_uri' => $secrets->apiBaseUri,
             'headers' => [
                 'Content-Type' => 'application/json',
                 'Authorization' => 'Bearer '.$this->fetchToken($feedSource),
@@ -163,8 +164,7 @@ class ApiClient
         } else {
             try {
                 $secrets = new SecretsDTO($feedSource);
-                $client = HttpClient::createForBaseUri($secrets->apiBaseUri);
-                $requestOptions = [
+                $response = $this->client->request('POST', $secrets->apiBaseUri.'/v1.0/generate-token', [
                     'headers' => [
                         'Content-Type' => 'application/json',
                         'Accept' => '*/*',
@@ -173,8 +173,7 @@ class ApiClient
                         'associationType' => self::AUTH_ASSOCIATION_TYPE,
                         'apiAuthKey' => $secrets->apiAuthKey,
                     ],
-                ];
-                $response = $client->request('POST', '/v1.0/generate-token', $requestOptions);
+                ]);
 
                 $content = $response->getContent(false); // Don't throw on non-2xx
                 $contentDecoded = json_decode($content, false, 512, JSON_THROW_ON_ERROR);