+++ tests/EventListener/RelationsChecksumListenerTest.php
@@ -617,8 +617,6 @@ class RelationsChecksumListenerTest extends KernelTestCase
 
         $playlistSlides = $playlist->getPlaylistSlides();
 
-        $this->assertGreaterThanOrEqual(10, $playlistSlides->count(), 'Fixtures count does not match expected value');
-
         $checksums = $playlist->getRelationsChecksum();
         $this->assertArrayHasKey('slides', $checksums);
         $slidesChecksum = $checksums['slides'];