+++ CHANGELOG.md
@@ -30,6 +30,10 @@ All notable changes to this project will be documented in this file.
 - Added relations checksum feature flag.
 - Fixes saving issues described in issue where saving resulted in infinite spinner.
 - Fixed loading of routes containing null string values.
+- Fixed relations checksum test.
+- Optimized release data fetching.
+- Optimized list loading.
+- Removed fixture length check from test.
 
 ### NB! Prior to 3.x the project was split into separate repositories